requirements.txt
glfw>=2.5
PyGLM>=2.7
numpy>=1.26
PyOpenGL>=3.1
PyOpenGL_accelerate>=3.1
httpx>=0.27
pydantic>=2.7
pillow>=10.0
moderngl>=5.10
# pygame-ce is the legacy prototype window layer — on the chopping block
# pygame-ce>=2.5